chore(package): update packages to fix breaking build
We had to regenerate the `yarn.lock` file to fix an issue with electron
and the download checksum. This regeneration introduced new issues with
the types and the PouchDB packages as the versions were updated. We now
use pinpointed versions for the PouchDB package and we don't need to
explicitly depend on `core-js` types.